refactor(geo): Make Spatial NLJ logic more explicit (#15088)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15088

To prepare for indexed spatial joins, restructure the SpatialJoinProbe
logic to:
1. Make the probeRow/buildVector/buildRow nested iteration more explicit
2. Move the output writing logic to a contained class to reduce action-at-distance
3. Allow writing output of more than one probeRow into one output vector

This will make the addition of spatial indices straightforward: Instead of
iterating over each buildRow, we only iterate over those rows returned by the
spatial index.

This also includes a significant number of VELOX_CHECKs to ensure invariants.
This will be helpful as we implement the spatial index.  They can be removed once
we have more confidence and if there are any significant performance implications.

No changes to behavior are expected.

Reviewed By: kgpai, xiaoxmeng

Differential Revision: D82970992

fbshipit-source-id: 4f8a952d0ba26e9c4ed622ac704f902af3fdd3b3