feat: Add Thrift service LocalRunnerService (#14390)
Summary:

Create Thrift service to execute Velox query plans locally. This service receives serialized query plans and uses LocalRunner to execute them. It then converts the Vector result as a thrift struct and sends it back to the caller. This will be used as part of our regression fuzzer to compare new changes to UDFs and expression evaluation execution with prior builds to identify behavior mismatches.

This diff contains:
* Header file with helper functions to convert vectors as thrift structs
* Cpp file that contains the service and runner logic
* Thrift class that contains the struct definitions
* Unit test file that tests mock request and response and vector conversion

Reviewed By: kagamiori

Differential Revision: D79850066