fix: error propagation in element_at UDF (#15110)
Summary:

Currently the UDF statically caches error messages to be re-used
whenever it needs to throw. However, since Velox error messages
add additional context when generated, the cached exception also
ends up re-using the message which contains the expression and
the top level expression details of the very first error produced
since process start. The original intention of that caching was
to mitigate perf impact when running under try where exceptions
are regularly expected. However, since we now have a mechanism
to handle that using Status objects, we employ that mechanism
to the UDF to fix the aforementioned issue.

Also updated the error messages to provide more details.
Updated unit tests and fixed a copy-paste error in one of the unit
tests which was testing the wrong input.

Differential Revision: D84308956