feat: Implement basic SpatialIndex (#15102)
Summary:

A SpatialIndex is a structure that when given a probe Envelope
will return the indices of potential build-side matches.

This implements a basic one: vectors for {min,max}{X,Y} and one for envelope
indices. This allows a quick binary search for the last minX that might match
the query envelope (query.maxX >= candidate.minX), potentially cutting off a
large number of candidates.  It then performance a vectorized linear searches
over the remaining vectors to assign any matching candidate indices.

There is also a pre-computed Envelope that contains all envelopes
in the index.  Checking against this first allows for short-circuiting.

The API for more advanced spatial indices should be the same, so this allows
us to hook up a spatial index to the join and later improve performance.

This index uses float32s instead of float64s.  This index is already
approximate, so a small number of false positives (coming from the difference
in 64- vs 32-bit precision) are acceptable.  The key is that the low-precision
envelopes must always be bigger than the full-precision envelopes, so we use
std::nextafter to bump the float32 in the right direction to make the envelopes
bigger.

Differential Revision: D84263281