fix(build): Patch required by macOS for setup abseil dependency
PR #14849 introduced abseil as regular dependency for re2
(and protobuf).
However, to build on macOS a patch is required to remove
-msse4.1 from the compile options.

This also addresses when changes to the setup files are made
but the macOS CI is not kicked off.
This would not have caught the issue because for macOS CI
almost all deps are bundled. For re2 this applied the patch.