fix: Make flatmap vector initialization more robust
Summary: The current logic assumes that the incoming vector is flat/array/map/row, hence doesn't apply additional checks. If this initialization is used in a place where various other encodings may be introduced, we may run into problems. This change added the check on desired encoding and resets the vector if encodings don't match

Differential Revision: D84537735