refactor: Add "Locked" suffix to thread-unsafe methods in AsyncDataCache (#15099)
Summary:
This aligns them with the project-wide convention for functions that
must be called under a lock, improving code consistency and making
locking requirements more explicit.

The `AsyncDataCacheEntry::movePromiseLocked()` method is only
accessible to `CacheShard` from outside the class. However, since
`CacheShard` is a friend of `AsyncDataCacheEntry`, it can be moved to
the `private` area to prevent others from accidentally calling this
thread-unsafe method.

No functional changes.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15099

Reviewed By: pratikpugalia

Differential Revision: D84365062

Pulled By: kKPulla

fbshipit-source-id: f1e3a63011b6e7ba1ca63edd27a7706c961112b9