fix: Make flatmap vector initialization more robust (#15152)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15152

The current logic assumes that the incoming vector is flat/array/map/row, hence doesn't apply additional checks. If this initialization is used in a place where various other encodings may be introduced, we may run into problems. This change added the check on desired encoding and resets the vector if encodings don't match

Reviewed By: Yuhta

Differential Revision: D84537735

fbshipit-source-id: 0cabb4b03fbc3bed93c4b860372f62a55dfe4f6d