fix: ConnectorSplit dataSource must be closed explicitly if present (#15197)
Summary:
`connectorSplit->dataSource` is an AsyncSource.
An AsyncSource must be explicitly closed or moved before being destroyed.
Internally, at IBM, we encountered a core dump in Presto C++ during a task cancellation and cleanup.
The `connectorSplit->dataSource` is set during `TableScan::preload` and might not be consumed if the task is cancelled.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15197

Reviewed By: tanjialiang

Differential Revision: D84858633

Pulled By: xiaoxmeng

fbshipit-source-id: cf3ddd07aa8b7ceb0c6abe1739d219d25d418d9f