feat(cudf): Add velox-cudf support for AssignUniqueId (#14766)
Summary:
This PR adds operator CudfAssignUniqueId to run AssignUniqueId in GPU.
Unit tests are added to verify the output, execution in GPU using stats.

The CudfAssignUniqueId offloads work to GPU with single call to list sequences function in libcudf. Other core logic of AssignUniqueId is same and the output of GPU operator should match exactly with CPU operator output.

Note on testing: Since the conversion operators will affect the batchsize to GPU operators, it needs to be set to minimal value for testing. Otherwise multiple batches are concatenated as single batch for GPU.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14766

Reviewed By: bikramSingh91

Differential Revision: D85060732

Pulled By: pedroerp

fbshipit-source-id: 072cee8b1dc2153e33c006fdedb4fed2b937db0a