feat(cudf): Upgrade cudf to 25.10 (#15183)
Summary:
Upgrade cudf to 25.10 and all dependent  libraries
This upgrade is required for newer features of kvikio.

This adds additional dependency in docker `libnvjitlink-devel-12-8` for CUDA 12.8.
For CUDA 13, it is already included.

When compiled with S3 enabled, conflicts arise due to linking to different libcurl.  kvikio compiles its own version of curl if system curl version is older. So, during linking, either one should be picked. Because aws-cpp-sdk-core cmake is hard coded to `/usr/lib54/libcurl.so`, it causes linking issues while linking with S3 and also cudf.
This has cmake workaround for the hardcoded linking of libcurl.so by aws-cpp-sdk-core, to replace it with soft link to libcurl.

Also add curl as dependency before S3 dependency is added so that system curl is not picked.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15183

Reviewed By: xiaoxmeng, peterenescu

Differential Revision: D85283050

Pulled By: mbasmanova

fbshipit-source-id: 67c668378ed7461e5b978e72331c1b10c576b430