refactor(string-view): Change makeInline() to take std::string_view (#15248)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15248

StringView::makeInline() used to take a std::string, which forced
users of the API to construct a local std::string when they don't already have
one. Changing to std::string_view. This shouldn't break backward compatibility
as std::string->std::string_view are implicitly convertible.

This is also safe because they API is only support for inline StringView, so
the contents are always copied.

Reviewed By: mbasmanova

Differential Revision: D85255859

fbshipit-source-id: 1f82bd661c396c9a03fe3b7c12ab10a5fd9582ca