fix: Remove redundant partition value handling in Iceberg column adaptation (#14516)
Summary:
Implement IcebergSplitReader::adaptColumns. Without overriding this method, current iceberg reader uses `SplitReader::adaptColumns` which is specific to Hive implementation. One difference between Hive and Iceberg is Iceberg spec requires all columns should be wrote to data files while in Hive, it only write non-partition columns to data file. So, there are special logic to handle this during read in `SplitReader::adaptColumns`. But in Iceberg this is not needed.
And for Hive, it populates the column data with partition value directly, this is correct for Hive only, but for Iceberg, there are different kinds transforms other than Identity transform. And we can not deduce the original column from the partition value.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14516

Reviewed By: Yuhta

Differential Revision: D85344487

Pulled By: mbasmanova

fbshipit-source-id: e0de6bdd44257a37f6727840803754639accdfd1