feat(geo): Increase bing tile max zoom shift to 6 (#15252)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15252

To protect against exponentially growing numbers of bing tiles,
we put a max zoom shift of 5.  This constrained a bing tile to be
shifted only 5 zoom levels, which increases the number by `4**5 == 2**10`.
When rolling out this feature, a user is getting errors because they were
using a shift of 6, previously allowed.

This makes the default 6, so one bing tile becomes `4**6 == 2**12`.

Reviewed By: kgpai, mbasmanova, Sullivan-Patrick

Differential Revision: D85278103

fbshipit-source-id: 56f02235734bd356e7a76821c562998fbf2a84b8