fix: Allow companion functions when result type is not resolvable given intermediate type (#11999)
Summary:
Allow generating companion functions even if they cannot be used standalone
with the understanding that these functions will be used as part of "single"
aggregations and return types will be provided explicitly.

The registrations of partial and merge companion functions does not require the
result type being resolvable given intermediate type. This PR removes the
limitations for them. The registration of merge_extract companion function used
to depend on the resolving for all aggregate steps, and this PR uses the passed-
in result type rather than the resolved one for single and final steps.

Decoupled the intermediate and result types resolving.

Details are in https://github.com/facebookincubator/velox/pull/11999#issuecomment-3274577979.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11999

Reviewed By: xiaoxmeng

Differential Revision: D84641699

Pulled By: kKPulla

fbshipit-source-id: 2e49251ec1acb518617d8d29e3de81580f821a2f