fix: Respect preferredOutputBatchBytes in StreamingAggregation (#15251)
Summary:

Today StreamingAggregation only looks at the number of groups (output rows) when deciding when to flush to the output Vector. This can lead to OOMs when the output rows are very wide, e.g. when using an aggregation that can grow very large depending on the input like ARRAY_AGG in Presto or COLLECT_LIST in Spark.

This change updates StreamingAggregation to look at the size in bytes of the rows in the RowContainer and compare that to preferredOutputBatchBytes as another trigger to output a Vector, allowing us to better control the memory usage and prevent OOMs.

Reviewed By: xiaoxmeng

Differential Revision: D85276743