fix: HashJoins and HashAggregations may produce incorrect results on types that support custom comparison (#15267)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15267

HashJoins and HashAggregations use HashTable to track values. The HashTable can use 3
different hash modes: hash, array, normalizedKey.  For the latter two, it relies on
VectorHasher to map values to a unique int64 (unlike hash this can't have collisions).
VectorHasher decides if it is able to do this based on the TypeKind (physical type).

This leads to bugs when the logical type supports custom comparison. In this case, the
TypeKind may indicate that value IDs can be produced, but the logic for doing this mapping
may map two different values which are equal using the Type's custom comparison to
different int64s at which point we lose the ability to identify equivalent values.

Furthermore, in HashJoin there is an additional issue with hash mode, where when it hits a
collision or does a lookup in the HashTable, it calls RowContainer's equals method which is
missing support for custom comparison. So values that should be equal so again we lose the
ability to identify equivalent values.

This change fixes these issues by:
1) In VectorHasher I changed typeSupportsValueIds to return false if the type supports
custom comparison
2) In VectorHasher I updated the constructor to explicitly disable range and unique value
(array/normalizedKey) based hashes if the type supports custom comparison.
3) I updated all functions in VectorHasher that rely on value IDs to do a DCHECK that the
type supports value IDs or handle them if possible. DCHECK because the code is
performance sensitive.
4) I removed the equals function from RowContainer. This was only used in one place in the
code and the logic is equivalent to compare == 0. It did have one optimization to skip null
checks if we know there are no nulls which I ported to compare.

Note that I removed the special handling for int128_t in VectorHasher's analyzeValue. That
was compensating for the issues I identify here in some cases, but did not fix the general
issue with types that are physically 128 bit integers being used as value IDs.  That fix is
coming separately.

Reviewed By: Yuhta, kgpai

Differential Revision: D85291290

fbshipit-source-id: 16ba2952324ffb887620489f7fd9686a90e7a9e6