build: Bump Presto version used in Fuzzer tests to 0.295 (#15334)
Summary:
This is to help address https://github.com/facebookincubator/velox/issues/15332

It looks like the version of Presto in the docker file currently being used is 0.290.  We need to build a new image, bumping the version kills two birds with one stone by triggering a new image to be built.

Also, fixed where we get the presto-cli executable jar, based on
https://github.com/prestodb/presto/issues/25976

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15334

Reviewed By: peterenescu, natashasehgal

Differential Revision: D85872823

Pulled By: kevinwilfong

fbshipit-source-id: ffe23d5394f1835f734032ca026d2815b2264bf0