refactor: Move map_subset registration logic into MapSubset.cpp
Summary:
The map_subset function in Presto has several optimized code paths for primitive types and varchars that complicate the registration logic.  I'd like to reuse the UDF in another engine with a different name, but I'd rather not have to maintain that logic in a separate location.

Given that the complicated registration is mostly due to implementation details of MapSubset itself, it makes sense to move it into MapSubset.cpp so that it can be shared wherever it is used. This is very common for other UDFs as well.

Differential Revision: D86125861