feat: Add HiveTableHandle::columnHandles (#15351)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15351

Currently if a partitioning column is not projected out, but used in remaining filter, the engine still adds it to assignments in order for the connector to recognize it.  This is a hack and we should pass the full column handles inside the table handle.

Fix https://github.com/facebookincubator/velox/issues/15348

Reviewed By: mbasmanova

Differential Revision: D85979627

fbshipit-source-id: f61ccbe848379a447c96e1f04276d30fb8665d1c