refactor: Deprecate exec::toSubfieldFilter API (#15392)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15392

Part of https://github.com/facebookincubator/velox/issues/15384

exec::toSubfieldFilter API appears language agnostic, but it assumes that function registered under name "not" has semantics of Presto's "not" function. This assumption is not valid and can lead to difficult to debug correctness issues.

The fix is to move that API to ExprToSubfieldFilterParser interface which has language-specific implementations.

Reviewed By: xiaoxmeng, Yuhta

Differential Revision: D86244462

fbshipit-source-id: 3ddf08ae81d2ef0b9cec96ff4e6d4588c7ded08c