feat: enable PartitionedOutputNode to distinguish the root of a plan
Currently when we translate a Presto plan to a Velox plan fragment, we
do not have a way later on at the Velox layer to detect if a
particular PartitionedOutputNode will be read by the coordinator as the
final output of a plan or will be read by other workers.

This patch adds a new field to PartitionedOutputNode so we can signal
this during plan translation inside Prestissimo.

In my particular case, This is needed to order to support exchange translation
decisions in the cuDF exchange (coming in via a later PR), where the
last PartitionedOutputNode has a single partition and will be read by
the Presto coordinator, but all other PartitionedOutputNodes will be
actually read by other workers via special, hardware accelerated
Exchange. Without this change we have no way to distinguish these two
at the pure Velox DriverAdapter translation level.