fuzzer: Adds an API to the vector fuzzer to corrupt data buffers for null rows
Summary:
This change introduces logic to deliberately corrupt the underlying
data for row indices that are marked as null in a vector. For
FlatVectors, the values buffer is corrupted at these null positions,
for Array and Map Vectors, the offsets and sizes are affected; and
for DictionaryVectors, the indices buffer is corrupted. For Constant
and RowVectors, this operation is a no-op. The purpose of this is to
generate test cases that ensure code correctly ignores data at null
positions, rather than relying on the assumption that the underlying
values are valid.

Differential Revision: D86273473