perf: Add Filter::testStringView
Summary: When testing filters against in-memory `StringView` vectors, we observe suboptimal performance when using the `testBytes` interface.  This is because `StringView` can compare the inline or prefix part in a few instructions before going into the byte-to-byte comparison loop, but `testBytes` does not use it.  Leverage this by adding a `testStringView` interface.  The gain is most significant for an equality filter with a small string (no longer than 12 bytes).

Differential Revision: D85272787