feat: Add rewrites for SWITCH and IF special forms (#15382)
Summary:
Rewrites SWITCH and IF special form expressions which have either of the shapes:
1. `SWITCH(condition1, value1, condition2, value2, .., conditionN, valueN, .., optional<defaultValue>)`
2. `IF(condition, trueValue, falseValue)`

Removes `(conditionN, valueN)` pairs from inputs when `conditionN` is false.
If `conditionN` is true and all the previous conditions were false, the expression is simplified to `valueN`.
If `conditionN` is true and there is at least one condition before `conditionN` that can not be evaluated without input data, modifies else clause to `valueN` and prunes rest of the inputs.
Returns an optimized SWITCH expression with pruned inputs.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15382

Reviewed By: kKPulla

Differential Revision: D86534408

Pulled By: peterenescu

fbshipit-source-id: ed12ad0d5bafb4e09e44e19ddc4b04f80648ab2f