feat: Remove cpr dependency and CMake variable leaks (#14160)
Summary:
cpr requires a specific libcurl version and is not portable. The current usage can be replaced by
libcurl which is present on most distros.
When cuDF is enabled, curl is bundled. Use the system or bundled curl otherwise for testing.
When a dependency uses `option` for a variable, it gets cached and becomes global.
However, if the variable is defined before, it does not get cached in some instances.
This PR unsets commonly used CMake variables BUILD_TESTING and BUILD_SHARED_LIBS if set by a dependency.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14160

Reviewed By: amitkdutta

Differential Revision: D86775547

Pulled By: peterenescu

fbshipit-source-id: f1a76ada3a86d47024f87b7b10a7e49092d002c9