fix(geo): Correct out-of-bounds index error in spatial join (#15472)
Summary:

We used candidateIndex_ (which indexes over all build side candidates)
into `decodedFilterResult_`, which just has rows for candidates for _this_
build vector. This caused array-out-of-bounds errors for large build sizes.
Similarily in `makeBuildVectorIndices`, we incorrectly indexed into rawIndices.

Our unit tests did not pick this up because they only tested batch sizes 1024 and 1.
1024 resulted in only one build side batch, and 1 meant that for vectors of size 1 default
initialization of indices to 0 meant that it was aways the right index. Adding a batch size of 2 find the indexing error.

Reviewed By: kgpai

Differential Revision: D86773251