refactor(geo): Calculate envelope bounds for spatial index outside of index (#15468)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15468

The RTreeSpatialIndex will require the overall bounds in order to calculate
the Hilbert index of each envelope. This can be done in the constructor, but will
require an additional iteration through all the envelopes. However, we've already
iterated through the envelopes when reading them, so we can construct the bounds
during that process, and pass it to the spatial index in the constructor.

This is more efficient, but it does create the possibility that SpatialJoinBuilder
calculates incorrect bounds.  We can guard against this by having some asserts in the
loop that constructs the spatial index if we want to make it a bit slower but safer.

Reviewed By: kgpai

Differential Revision: D86127951

fbshipit-source-id: 0bed833c9c7635052af261358badf0a5a122894d