feat: Add rewrite for COALESCE special form (#15399)
Summary:
Rewrites `COALESCE` special form expression. First, flattens nested `COALESCE` expressions in a recursive manner, `eg: COALESCE(a, COALESCE(b, c)) -> COALESCE(a, b, c)`.

The inputs to `COALESCE` are then traversed in order and:
1. If the input is constant:
a. and it is `NULL`, the input is removed.
b. and if there are no inputs before this, the `COALESCE` expression is simplified to this constant input.
c. otherwise, all subsequent inputs after this are removed.
2. Otherwise, if the same input is seen before, it is removed i.e non-constant inputs to `COALESCE` are deduplicated.

If there is a single input to `COALESCE` after pruning inputs, `COALESCE` is simplified to this expression.
Otherwise, returns an optimized `COALESCE` expression with pruned inputs.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15399

Reviewed By: bikramSingh91

Differential Revision: D86775529

Pulled By: peterenescu

fbshipit-source-id: dcb076a98fbcdf388e5eba595e7dd4d57b612135