fix: FlatVector::resizeValues can lose existing values for opaque types
Summary:
There is a bug in FlatVector::resizeValues where for non-POD types (I think this is just opaque 
types) if the Buffer is immutable, there's a typo where it copies values from the new Buffer
into itself, losing the existing values.

Fixing this bug exposes another bug that FlatVector::resizeValues uses BaseVector::length_
as the previous size, however, by this point BaseVector::resize has already been invoked, so it
should be equal to new size (leading to reading off the end of the Buffer after the fix for the
first issue).

To address this I pass the previous size to FlatVector::resizeValues as relying on the order of 
invoking this function and BaseVector::resize seems error prone.

Another option would be to rely on the Buffer sizes, but then we may be copying more values
than necessary. Along these lines I updated the logic for POD types (other types) to use the
Vector size to determine how many values to copy over (any other values would have been set
to initialValue already if one is needed).

I noticed a similar bug in the bool template specialization of FlatVector::resizeValues where it
was using BaseVector::length_ for deciding how many values to set to initialValue if we are
just updating the size of an existing Buffer with sufficient capacity. I opted to just delete this
code since it is in a private function and there is no invocation that passes an initialValue for a
bool, and it's inconsistent with the base implementation of FlatVector::resizeValues which only
sets values to initialValue if the Buffer is reallocated.

Differential Revision: D86920858