fix: Fix StringView remapping in transferOrCopyTo() when string buffers are copied (#15489)
Summary:

FlatVector::transferOrCopyTo() and ConstantVector::transferOrCopyTo() may copy the string 
buffers to new ones. However, there was a bug that they didn't udpate the StringViews in the 
vectors afterwards. StringViews must be updated if a string buffer is replaced by its copy, 
because StringViews contain pointers to addresses inside the string buffers.

This diff updates StringViews in FlatVector::transferOrCopyTo() and 
ConstantVector::transferOrCopyTo() when a copy of a string buffer occurs. It also updated the 
unit test to cover these situations. 

In addition, this diff adds documentation of the BaseVector::transferOrCopyTo() API. It also 
removes FlatMapVector::transferOrCopyTo() since it's not covered in the unit test yet.

Reviewed By: bikramSingh91

Differential Revision: D86680882