fix: Don't corrupt the FlatVector if resize fails to allocate memory (#15495)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15495

FlatVector::resize currently calls BaseVector::resize first then resizes the values_ Buffer. If we
fail to resize the values_ Buffer (e.g. we run out of memory, or the Driver is shutting down) it
can leave the FlatVector in an invalid state where values_ isn't large enough to hold values_
elements.

This change modifies the logic to only call BaseVector::resize after resizing the values_ Buffer.
This way if we fail at worst we have a values_ Buffer that is larger than length_ which is a valid
state to be in. Note that BaseVector::resize already allocates the nulls Buffer before updating
length_. Also DictionaryVector already allocates the indices Buffer before calling
BaseVector::resize, and ConstantVector doesn't resize Buffers.

I have validated that after https://github.com/facebookincubator/velox/pull/15487 we no longer
refer to any values from BaseVector that may be modified by resize (e.g. length_) when
resizing the values_ Buffer.

When we can't allocate the Buffer we throw a VeloxRuntimeError, and we make no guarantees
about the validity of objects within the scope of the Driver after a VeloxRuntimeError is
thrown, so this is just a best effort attempt to reduce the risk of undefined behavior when
users attempt to handle VeloxRuntimeErrors incorrectly (e.g. in UDFs).

Reviewed By: bikramSingh91

Differential Revision: D86994206

fbshipit-source-id: fbc525b66fb68caefe7bc410427f96f9d61f02c6