build(ci): Run clang-tidy and add license header annotations (#15284)
Summary:
It might seems like overkill to use the adapters image but for clang-tidy to run it needs `compile-commands.json`, so we need to run cmake, which needs the dependencies.

I use `--line-filter` to only output warnings on changed lines, other warnings would not show up as annotations anyway and take very long.

The clang-tidy job only runs when there are changes to C++ files.

Closes: https://github.com/facebookincubator/velox/issues/15047
Closes: https://github.com/facebookincubator/velox/issues/15050

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15284

Reviewed By: kKPulla

Differential Revision: D87232092

Pulled By: Yuhta

fbshipit-source-id: dacf3e3f6bb67f09e87a4793229240b38c718118