fix: Ensure the MapWriter/ArrayWriter doesn't end up in an inconsistent state when resizing children fails (#15497)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15497

MapWriter and ArrayWriter both update their capacity_ before resizing their children writers. If
we fail to resize the writers (e.g. we run out of memory, or the Driver is shutting down) it
can leave the Writer in an invalid state where capacity_ is larger than the size of the Vectors in
the children.

This change modifies the logic to only update capacity_ after resizing the children. This way if
we fail at worst we have a child whose size is greater than capacity_ which is a valid state to
be in.

When we fail to resize the child we typically throw a VeloxRuntimeError (e.g. when trying to
allocate memory), and we make no guarantees about the validity of objects within the scope
of the Driver after a VeloxRuntimeError is thrown, so this is just a best effort attempt to reduce
the risk of undefined behavior when users attempt to handle VeloxRuntimeErrors incorrectly
(e.g. in UDFs).

Reviewed By: bikramSingh91

Differential Revision: D87005822

fbshipit-source-id: 29f1eb07cee8552b1161f58ecfd11cfb9dcf03a0