refactor(exception): Decouple line number logging logic (#15521)
Summary:

# Rationale

Just decoupling the logic to append a number into a string.

Apparently, the code does not want to allocate any extra memory for the operation. I'd like to make the code structure abstract the implementation detail. I added a small comment for future readers that I had to figure out by spending time to understand.

I hope this is reasonable.

Differential Revision: D87240823