refactor: Mark ExprToSubfieldFilterParser::toSubfieldFilter API as test-only (#15542)
Summary:

ExprToSubfieldFilterParser::toSubfieldFilter is used only in PlanBuilder and in velox/exec/tests/FilterToExpressionTest.cpp. 

It would be nice to move this API into PlanBuilder and remove its usage from FilterToExpressionTest.

The implementation had a bug where it would drop disjuncts from an OR that has > 2 inputs. Added a check to fail in this case instead of producing wrong results.

Reviewed By: amitkdutta, xiaoxmeng

Differential Revision: D87343576