feat: Support matchRowFieldsByName in CastHooks
Summary:
Currently, CastExpr determines whether to map row fields by name or by position when
casting between row types, based on a setting in the QueryConfig. This means that it must
apply uniformly to all CastExprs in a query.

We have a use case in Spark where the engine supports cast mapping row fields by position
by default, but a custom UDF enables casting mapping row fields by name. To enable this I'm
adding a function to the CastHooks interface that we check to determine how to map
fields. For PrestoCastHooks this defaults to the value from QueryConfig so there's no change
in behavior. For SparkCastHooks it looks at the value from the QueryConfig as well as a
constructor parameter and OR's them, preserving the existing behavior unless we explicitly
set the new behavior when constructing the instance of SparkCastHooks.

I changed the name in the CastHooks to matchRowFieldsByName instead of 
matchStructByName because we're matching the fields not the struct, and we refer to the
type as row in Velox not struct. Note that matchStructByName is hardcoded in the config
name string, so to not break backwards compatibility I did not change it there.

Differential Revision: D87371309