fix: Max length estimation in DateTimeFormatter (#15532)
Summary:

Fixes the maxLength() utility function in DateTimeFormatter for TIMEZONE specifiers that print out the timezone abbreviation, that is, for 'z', 'zz' and 'zzz'. It increases the size for the case where some timezones do not have abbreviations but instead use GMT offsets like `GMT+05:00`.

Originally found by the expression fuzzer.

Reviewed By: peterenescu

Differential Revision: D87290122