perf: Use FMT_COMPILE for float to string casting (#15518)
Summary:
Make the float to string casts use FMT_COMPILE, to avoid parsing the format string on every call. Also take a reference to the expected result, to avoid copying the string on every cast.

Together, these improve the performance of the basic benchmarks for real and double to scientific/standard notation casts.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15518

Reviewed By: bikramSingh91

Differential Revision: D87352405

Pulled By: Yuhta

fbshipit-source-id: a5e640da8f3a3fb1a09744fd37fd947df7ee065b