feat: Add current_timezone function (#14739)
Summary:
In the past, we were not implementing functions like current_timezone that returned constants in Velox. This was because Presto Java would infer the constant value at the co-ordinator and fold it, so the worker was never passed this function in execution.

However, with recent changes for side-car and worker based constant folding implementation, we need native implementations for functions like current_timezone, so adding to Velox as well.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14739

Reviewed By: pratikpugalia, kagamiori

Differential Revision: D87390791

Pulled By: kgpai

fbshipit-source-id: a48524a94cc1fc7b1c9fc075981d160c45c6c0e1