fix: Improve serde behavior for zero-area polygons (#15558)
Summary:

Today there is an inconsistency Java and C++ with polygon serde related to orientation. A polygon ring's orientation, that is, if the edges run clockwise(CW) or counterclockwise(CCW), is sometimes used to determine if the polygon is an exterior ring (shell) or interior ring (hole). During Esri Java and C++ serialization, we try to 'canonicalize' orientation of polygons to fit this rule. Canonicalization means reversing any shells that are CCW or holes that are CW. However, some polygons have exactly zero area, which are considered CCW before and after reversal. This causes problems after serialization and deserialization, which relies on orientation to determine if a ring is a shell or hole: zero-area polygon shells can be misinterpreted as holes.

This issue is further exacerbated due to Java having 2 serde paths, with C++ serde mirroring the JTS serde path. Esri serde path in Java does not have the same polygon orientation checks. This results in the following behavior:

Fails with `shell is empty but holes are not` in Java:
```
select st_area(st_geometryfromtext('POLYGON((-135 85, -45 85, 45 85, 135 85, -135 85))'))
```

Succeeds in Java:
```
select st_area(to_spherical_geography(st_geometryfromtext('POLYGON((-135 85, -45 85, 45 85, 135 85, -135 85))')))
```

Succeeds in Java, but incorrectly reconstructs the MultiPolygon as `MULTIPOLYGON (((1 1, 2 2, 2 1, 1 1), (1 1, 2 2, 3 3, 2 2, 1 1)))`:
```
select st_astext(st_geometryfromtext('MULTIPOLYGON (((1 1, 2 1, 2 2, 1 1)), ((1 1, 2 2, 3 3, 2 2, 1 1)))'))
```

After this change:

-For polygons which have zero area shells or holes, we simply leave them alone. The first ring is *always* a shell, all further rings are *always* holes. This means no more failing with `shells are empty but holes are not`

-We no longer reverse all zero-area shells. Previously, zero-area shells would try to 'canonicalize' and reverse the shell, leaving a (still zero-area) reversed orientation. Serde will leave zero-area shells alone.

-For MultiPolygons, we *always* fail when encountering zero-area rings. We could theoretically allow zero-area holes, open to discussion here. But I think it's a simpler contract if MultiPolygons can never have any zero-area rings.

Reviewed By: jagill, kgpai

Differential Revision: D87384761