perf: Optimize copying distinct spilling vector to output vector in HashAggragtion (#15583)
Summary:
Currently, when it comes to copying a distinct spilling vector to the output
vector, the `GroupingSet::mergeNextWithoutAggregates` method leverages
`RowVector::copy`. However, this approach is not efficient, as it involves
operations which are not necessary for this scenario such as allocating the
source to the target index buffer, ensuring that the output vector is writable,
and copying ascii-ness info. This PR utilizes `OperatorUtils::gatherCopy`
instead.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15583

Reviewed By: kgpai

Differential Revision: D87666447

Pulled By: kagamiori

fbshipit-source-id: 8eee4968edeb402c3064a00188d55f84b78b51b2