feat(vector): Add retainedSize(totalStringBufferSize) API to return total size of all string buffers (#15578)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15578

This diff adds a new API `uint64_t retainedSize(uint64_t& totalStringBufferSize)` in BaseVector
to return total size of all string buffers recursively carried by the vector through the out
parameter, in addition to the total size of all buffers.

This is needed when the string buffers are multiply referenced and we want to amortize the size,
such as in the buffer mode in LocalPartition.

bypass-github-export-checks

Reviewed By: mbasmanova

Differential Revision: D87505861

fbshipit-source-id: 5b950a8819f9c244b65f71a627a1bcdce97c0966